reading a size for task

+++ server/models/Task.js
@@ -18,7 +18,7 @@ module.exports = (sequelize, DataTypes) => {
         },
         swimlaneOrderNumber: DataTypes.INTEGER,
         color: DataTypes.STRING,
-        size: DataTypes.INTEGER,
+        size: DataTypes.DOUBLE,
         difficulty: DataTypes.INTEGER,
     })
     Task.associate = (models) => {
